Create ability to face a direction at a tap.
Used @window's timer to accomplish this; everytime the directional input is calculated, it sets a short timer for 5 ticks (1/12 second) before reading any more input. This creates a very slight but barely noticeable delay when a player is holding down a directional button, but allows for a light tap to just change facing direction. Note this delay is ONLY set if the player is NOT already facing that direction.

Also did a bit of cleaning up in relevant methods, mostly deleting unused code.